rag-backend/requirements.txt
fastapi
uvicorn
python-dotenv
qdrant-client
psycopg2-binary
openai
tiktoken
markdown
